// src/api/model/internalHandlerWSCommandDoc.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * 故桌 Narratable API
 * 故桌（Narratable）实时跑团平台的后台 API
 * OpenAPI spec version: 1.0
 */
import type { GithubComNarraTableBackendPkgProtocolBlock } from './githubComNarraTableBackendPkgProtocolBlock.ts';
import type { GithubComNarraTableBackendPkgProtocolClueContent } from './githubComNarraTableBackendPkgProtocolClueContent.ts';
import type { GithubComNarraTableBackendPkgProtocolForwardContent } from './githubComNarraTableBackendPkgProtocolForwardContent.ts';
import type { GithubComNarraTableBackendPkgProtocolMessageType } from './githubComNarraTableBackendPkgProtocolMessageType.ts';
import type { GithubComNarraTableBackendPkgProtocolParam } from './githubComNarraTableBackendPkgProtocolParam.ts';
import type { GithubComNarraTableBackendPkgProtocolVeil } from './githubComNarraTableBackendPkgProtocolVeil.ts';
import type { GithubComNarraTableBackendPkgProtocolWSCommandType } from './githubComNarraTableBackendPkgProtocolWSCommandType.ts';

export interface InternalHandlerWSCommandDoc {
  accessToken?: string;
  clue?: GithubComNarraTableBackendPkgProtocolClueContent;
  command?: string;
  content?: GithubComNarraTableBackendPkgProtocolBlock[];
  eventSeq?: number;
  forward?: GithubComNarraTableBackendPkgProtocolForwardContent;
  instanceId?: string;
  messageType?: GithubComNarraTableBackendPkgProtocolMessageType;
  params?: GithubComNarraTableBackendPkgProtocolParam[];
  raw?: string;
  requestId?: string;
  roomId?: string;
  senderMaskId?: string;
  type?: GithubComNarraTableBackendPkgProtocolWSCommandType;
  veil?: GithubComNarraTableBackendPkgProtocolVeil;
}