// src/api/model/internalHandlerWSEventDoc.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * 故桌 Narratable API
 * 故桌（Narratable）实时跑团平台的后台 API
 * OpenAPI spec version: 1.0
 */
import type { GithubComNarraTableBackendPkgProtocolAuthResult } from './githubComNarraTableBackendPkgProtocolAuthResult.ts';
import type { GithubComNarraTableBackendPkgProtocolEventType } from './githubComNarraTableBackendPkgProtocolEventType.ts';
import type { GithubComNarraTableBackendPkgProtocolMessage } from './githubComNarraTableBackendPkgProtocolMessage.ts';
import type { GithubComNarraTableBackendPkgProtocolPresence } from './githubComNarraTableBackendPkgProtocolPresence.ts';
import type { GithubComNarraTableBackendPkgProtocolWSError } from './githubComNarraTableBackendPkgProtocolWSError.ts';

export interface InternalHandlerWSEventDoc {
  auth?: GithubComNarraTableBackendPkgProtocolAuthResult;
  error?: GithubComNarraTableBackendPkgProtocolWSError;
  eventSeq?: number;
  gap?: boolean;
  instanceId?: string;
  /** message.created / message.updated */
  message?: GithubComNarraTableBackendPkgProtocolMessage;
  /** message.deleted */
  messageId?: string;
  /** veil.revealed */
  messageIds?: string[];
  presence?: GithubComNarraTableBackendPkgProtocolPresence;
  /** "space_closed" */
  reason?: string;
  replay?: boolean;
  requestId?: string;
  roomId?: string;
  spaceId?: string;
  timestamp?: string;
  type?: GithubComNarraTableBackendPkgProtocolEventType;
}